powertop: don't look for kbd input in dump mode
When in dump mode, don't check for keyboard input (i.e., don't
check fd0 in the select() syscall).  Otherwise when 'key' is non-zero,
ticktime (from -t) is destroyed and the program loop exits
quickly.  If someone wants to kill powertop when in dump mode,
just kill it with ^C or "kill".

Signed-off-by: Randy Dunlap <randy.dunlap@oracle.com>